+++ Drivers/zooz/zooz-zen-dimmer.groovy
@@ -1,11 +1,22 @@
 /*
  *  Zooz ZEN Dimmers Universal
- *    - Model: ZEN22/24 - MINIMUM FIRMWARE 3.07
+ *    - Model: ZEN22, ZEN24 - MINIMUM FIRMWARE 3.07
  *    - Model: ZEN27 - MINIMUM FIRMWARE 2.08
- *    - Model: ZEN72
- *    - Model: ZEN77
+ *    - Model: ZEN72, ZEN74, ZEN77 - All Firmware
  *
  *  Changelog:
+
+## [1.4.3] - 2021-04-21 (@jtp10181)
+  ### Added
+  - ZEN30 Uses new custom child driver by default, falls back to hubitat generic
+  - Command to change indicator on/off settings
+  - Support for ZEN73 and ZEN74
+  - Support for Push, Hold, and Release commands
+  ### Changed
+  - Removed unnecessary capabilities
+  - Renamed indicatorColor to setLED to match other Zooz drivers
+  ### Fixed
+  - Status Syncing... was not always updating properly
 
 ## [1.4.2] - 2021-01-31 (@jtp10181)
   ### Added
@@ -159,8 +170,10 @@ CommandClassReport- class:0x9F, version:1
 @Field static final int maxAssocGroups = 3
 @Field static final int maxAssocNodes = 5
 
-@Field static Map deviceModelNames = ["B111:1E1C":"ZEN21", "B112:1F1C":"ZEN22", "B111:251C":"ZEN23", "B112:261C":"ZEN24", 
-	"A000:A001":"ZEN26", "A000:A002":"ZEN27", "7000:A001":"ZEN71", "7000:A002":"ZEN72", "7000:A006":"ZEN76", "7000:A007":"ZEN77"]
+@Field static Map deviceModelNames =
+	["B111:1E1C":"ZEN21", "B112:1F1C":"ZEN22", "B111:251C":"ZEN23", "B112:261C":"ZEN24",
+	"A000:A001":"ZEN26", "A000:A002":"ZEN27", "7000:A001":"ZEN71", "7000:A002":"ZEN72",
+	"7000:A003":"ZEN73", "7000:A004":"ZEN74", "7000:A006":"ZEN76", "7000:A007":"ZEN77"]
 
 @Field static Map paddleControlOptions = [0:"Normal", 1:"Reverse", 2:"Toggle Mode"]
 @Field static Map ledModeOptions = [0:"LED On When Switch Off", 1:"LED On When Switch On", 2:"LED Always Off", 3:"LED Always On"]
@@ -181,11 +194,11 @@ CommandClassReport- class:0x9F, version:1
 @Field static Map relayBehaviorOptions = [0:"Reports Status & Changes LED Always", 1:"Doesn't Report Status or Change LED"]
 @Field static Map zwaveRampRateOptions = [0:"Match Physical Ramp Rate", 1:"Z-Wave Can Set Ramp Rate [RECOMMENDED]"]
 @Field static Map associationReportsOptions = [
-	0:"None", 1:"Physical Tap On ZEN Only", 2:"Physical Tap On Connected 3-Way Switch Only", 3:"Physical Tap On ZEN / 3-Way Switch", 
-	4:"Z-Wave Command From Hub", 5:"Physical Tap On ZEN / Z-Wave Command", 6:"Physical Tap On 3-Way Switch / Z-Wave Command", 
-	7:"Physical Tap On ZEN / 3-Way Switch / Z-Wave Command", 8:"Timer Only", 9:"Physical Tap On ZEN / Timer", 
-	10:"Physical Tap On 3-Way Switch / Timer", 11:"Physical Tap On ZEN / 3-Way Switch / Timer", 12:"Z-Wave Command From Hub / Timer", 
-	13:"Physical Tap On ZEN / Z-Wave Command / Timer", 14:"Physical Tap On ZEN / 3-Way Switch / Z-Wave Command / Timer", 
+	0:"None", 1:"Physical Tap On ZEN Only", 2:"Physical Tap On Connected 3-Way Switch Only", 3:"Physical Tap On ZEN / 3-Way Switch",
+	4:"Z-Wave Command From Hub", 5:"Physical Tap On ZEN / Z-Wave Command", 6:"Physical Tap On 3-Way Switch / Z-Wave Command",
+	7:"Physical Tap On ZEN / 3-Way Switch / Z-Wave Command", 8:"Timer Only", 9:"Physical Tap On ZEN / Timer",
+	10:"Physical Tap On 3-Way Switch / Timer", 11:"Physical Tap On ZEN / 3-Way Switch / Timer", 12:"Z-Wave Command From Hub / Timer",
+	13:"Physical Tap On ZEN / Z-Wave Command / Timer", 14:"Physical Tap On ZEN / 3-Way Switch / Z-Wave Command / Timer",
 	15:"All Of The Above" ]
 
 metadata {
@@ -196,30 +209,31 @@ metadata {
 		importUrl: "https://raw.githubusercontent.com/jtp10181/hubitat/master/Drivers/zooz/zooz-zen-dimmer.groovy"
 	) {
 		capability "Actuator"
-		capability "Sensor"
 		capability "Switch"
 		capability "SwitchLevel"
-		//capability "Light"  //Redundant with Switch
 		capability "Configuration"
 		capability "Refresh"
-		capability "HealthCheck"
 		capability "PushableButton"
 		capability "HoldableButton"
 		capability "ReleasableButton"
-		capability "DoubleTapableButton"
+		//capability "DoubleTapableButton"
 
-		command "paramCommands", [[name:"Select One*", type: "ENUM", constraints: ["Refresh","Test All","Hide Invalid","Clear Hidden"] ]]
-		command "indicatorColor", [[name:"Works ONLY on ZEN7x Series!*", type: "ENUM", constraints: ledColorOptions ]]
+		command "paramCommands", [[name:"Select Command*", type: "ENUM", constraints: ["Refresh","Test All","Hide Invalid","Clear Hidden"] ]]
+		command "setLED", [
+			[name:"Select Color*", description:"Works ONLY on ZEN7x Series!", type: "ENUM", constraints: ledColorOptions] ]
+		command "setLEDMode", [
+			[name:"Select Mode*", description:"This Sets Preference (#2)*", type: "ENUM", constraints: ["Default","Reverse","Off","On"]] ]
 
 		attribute "assocDNI2", "string"
 		attribute "assocDNI3", "string"
 		attribute "syncStatus", "string"
 
-		fingerprint mfr:"027A", prod:"B112", deviceId:"1F1C", inClusters:"0x5E,0x6C,0x55,0x9F", deviceJoinName:"Zooz ZEN22 Dimmer"
-		fingerprint mfr:"027A", prod:"B112", deviceId:"261C", inClusters:"0x5E,0x6C,0x55,0x9F", deviceJoinName:"Zooz ZEN24 Dimmer"
-		fingerprint mfr:"027A", prod:"A000", deviceId:"A002", inClusters:"0x5E,0x6C,0x55,0x9F", deviceJoinName:"Zooz ZEN27 S2 Dimmer"
-		fingerprint mfr:"027A", prod:"7000", deviceId:"A002", inClusters:"0x5E,0x6C,0x55,0x9F", deviceJoinName:"Zooz ZEN72 Switch"
-		fingerprint mfr:"027A", prod:"7000", deviceId:"A007", inClusters:"0x5E,0x6C,0x55,0x9F", deviceJoinName:"Zooz ZEN77 S2 Dimmer"
+		fingerprint mfr:"027A", prod:"B112", deviceId:"1F1C", deviceJoinName:"Zooz ZEN22 Dimmer"
+		fingerprint mfr:"027A", prod:"B112", deviceId:"261C", deviceJoinName:"Zooz ZEN24 Dimmer"
+		fingerprint mfr:"027A", prod:"A000", deviceId:"A002", deviceJoinName:"Zooz ZEN27 S2 Dimmer"
+		fingerprint mfr:"027A", prod:"7000", deviceId:"A002", deviceJoinName:"Zooz ZEN72 Dimmer"
+		fingerprint mfr:"027A", prod:"7000", deviceId:"A004", deviceJoinName:"Zooz ZEN74 Dimmer"
+		fingerprint mfr:"027A", prod:"7000", deviceId:"A007", deviceJoinName:"Zooz ZEN77 S2 Dimmer"
 	}
 
 	preferences {
@@ -269,6 +283,18 @@ String getAssocDNIsSetting(grp) {
 	return ((val && (val.trim() != "0")) ? val : "") 
 }
 
+void push(buttonId) { sendBasicButtonEvent(buttonId, "pushed") }
+void hold(buttonId) { sendBasicButtonEvent(buttonId, "held") }
+void release(buttonId) { sendBasicButtonEvent(buttonId, "released") }
+void doubleTap(buttonId) { sendBasicButtonEvent(buttonId, "doubleTapped") }
+
+void sendBasicButtonEvent(BigDecimal buttonId, String name) {
+	Map event = [name: name, value: buttonId, type:"digital", isStateChange:true]
+	event.descriptionText="button ${buttonId} ${name}"
+	logTxt "${event.descriptionText} (${event.type})"
+	sendEvent(event)
+}
+
 void paramCommands(str) {
 	switch (str) {
 		case "Refresh":
@@ -292,7 +318,7 @@ void paramCommands(str) {
 void paramsTestAll() {
 	Map configsMap = getParamStoredMap()
 	List lastTest = state.tmpLastTest.collect()
-	Integer key = configsMap.find { !lastTest || it.key > lastTest[0] } ?.key
+	Integer key = configsMap.find{ !lastTest || it.key > lastTest[0] }?.key
 	if (!key) {
 		logDebug "Finished Testing All Params"
 		runInMillis(1400, paramsHideInvalid)
@@ -336,7 +362,7 @@ void paramsHideInvalid() {
 		//Clean up configVals, remove hidden params
 		configDisabled.each { configsMap.remove(it) }
 		device.updateDataValue("configVals", configsMap.inspect())
-		runIn(1,refreshSyncStatus)
+		updateSyncingStatus()
 	}
 	else {
 		logDebug "Disabled Parameters: NONE"
@@ -353,20 +379,38 @@ void paramsClearHidden() {
 	state.remove("tmpLastTest")
 	state.remove("tmpFailedTest")
 	device.removeDataValue("configHide")
-	runIn(1,refreshSyncStatus)
+	updateSyncingStatus()
 
 	sendEvent(name: "WARNING", value: "COMPLETE - RELOAD THE PAGE!", isStateChange: true)
 }
 
-void indicatorColor(color) {
+void setLED(String colorName) {
 	if (!(state.deviceModel ==~ /ZEN7\d/)) {
 		log.warn "Indicator Color can only be changed on ZEN7x models"
 		return
 	}
 
-	def paramVal = ledColorOptions.find{it.value == color}?.key
-	logDebug "indicatorColor Value: ${color} : ${paramVal}"
-	sendCommands(configSetGetCmd(ledColorParam, paramVal))
+	def param = ledColorParam
+	Integer paramVal = ledColorOptions.find{ it.value?.toUpperCase() == colorName?.toUpperCase() }?.key
+	logDebug "Indicator Color Value [${colorName} : ${paramVal}]"
+	//Set the Preference to match new setting, then send command to device
+	device.updateSetting("configParam${param.num}",[value:"${paramVal}", type:"enum"])
+	sendCommands(configSetGetCmd(param, paramVal))
+}
+
+void setLEDMode(String modeName) {
+	if (deviceModelShort in [23,24]) {
+		log.warn "There is No Indicator on ZEN23/24 models"
+		return
+	}
+
+	def param = ledModeParam
+	Map modeMap = ["default":0,"reverse":1,"off":2,"on":3]
+	Integer paramVal = modeMap[modeName?.toLowerCase()] ?: 0
+	logDebug "Indicator Value [${modeName} : ${paramVal}]"
+	//Set the Preference to match new setting, then send command to device
+	device.updateSetting("configParam${param.num}",[value:"${paramVal}", type:"enum"])
+	sendCommands(configSetGetCmd(param, paramVal))
 }
 
 
@@ -388,17 +432,11 @@ def updated() {
 
 		initialize()
 
-		runIn(2, executeConfigureCmds, [overwrite: true])
+		runIn(1, executeConfigureCmds, [overwrite: true])
 	}
 }
 
 void initialize() {
-	def checkInterval = ((60 * 60 * 3) + (5 * 60))
-
-	if (!device.currentValue("checkInterval")) {
-		sendEvent(name: "checkInterval", value:checkInterval, displayed:false, data:[protocol: "zwave", hubHardwareId: device.hub.hardwareID, offlinePingable: "1"])
-	}
-
 	if (!device.currentValue("numberOfButtons")) {
 		sendEvent(name:"numberOfButtons", value:10, displayed:false)
 	}
@@ -416,7 +454,9 @@ def configure() {
 		state.resyncAll = true
 	}
 
+	updateSyncingStatus()
 	runIn(2, executeRefreshCmds, [overwrite: true])
+	runIn(5, updateSyncingStatus, [overwrite: true])
 	runIn(8, executeConfigureCmds, [overwrite: true])
 }
 
@@ -431,7 +471,7 @@ void executeConfigureCmds() {
 		cmds << switchMultilevelGetCmd()
 	}
 
-	if (state.resyncAll || !state.deviceModel || !firmwareVersion) {
+	if (state.resyncAll || !firmwareVersion || !state.deviceModel) {
 		cmds << versionGetCmd()
 	}
 
@@ -565,10 +605,8 @@ private getAssocDNIsSettingNodeIds(grp) {
 	return nodeIds
 }
 
-
-def ping() {
-	logDebug "ping..."
-	return switchMultilevelGetCmd()
+private getDeviceModelShort() {
+	return safeToInt(state?.deviceModel.drop(3))
 }
 
 
@@ -614,7 +652,7 @@ def refresh() {
 }
 
 void executeRefreshCmds() {
-	refreshSyncStatus()
+	updateSyncingStatus()
 
 	List<String> cmds = []
 	cmds << versionGetCmd()
@@ -624,7 +662,7 @@ void executeRefreshCmds() {
 }
 
 void paramsRefresh() {
-	refreshSyncStatus()
+	updateSyncingStatus()
 
 	List<String> cmds = []
 	configParams.each { param ->
@@ -1035,12 +1073,12 @@ List<Map> getConfigParams() {
 	}	
 
 	// ZEN23/24 does not have a LED at all
-	if (state.deviceModel in ["ZEN23", "ZEN24"]) {
+	if (deviceModelShort in [23,24]) {
 		params.removeAll { it == ledModeParam }
 	}
 
 	// Remove from all except ZEN21/22/23/24/71/72
-	if (!(state.deviceModel ==~ /ZEN\d[1234]/)) {
+	if (!(deviceModelShort in [21,22,23,24,71,72])) {
 		params.removeAll { it == threeWaySwitchTypeParam }
 	}
 
@@ -1071,7 +1109,9 @@ Map getPaddleControlParam() {
 }
 
 Map getLedModeParam() {
-	return getParam(2, "LED Indicator", 1, 0, ledModeOptions)
+	// ZEN73/74 have LED but it is hidden and defaulted to always off
+	Integer defaultVal = (deviceModelShort in [73,74]) ? 2 : 0
+	return getParam(2, "LED Indicator", 1, defaultVal, ledModeOptions)
 }
 
 Map getAutoOffEnabledParam() {
@@ -1121,7 +1161,7 @@ Map getMaximumBrightnessParam() {
 }
 
 Map getDoubleTapBrightnessParam() {
-	Map options = (state.deviceModel in ["ZEN77"]) ? doubleTapBrightness700Options : doubleTapBrightnessOptions
+	Map options = (state.deviceModel ==~ /ZEN7\d/) ? doubleTapBrightness700Options : doubleTapBrightnessOptions
 	return getParam(12, "Double Tap Up Brightness", 1, 0, options)
 }
 
@@ -1157,7 +1197,7 @@ Map getCustomBrightnessParam() {
 
 // ZEN21/22/23/24/71/72 Only
 Map getThreeWaySwitchTypeParam() {
-	// All the same for dimmers
+	// All the same number for dimmers (19)
 	Integer num = 19
 	return getParam(num, "3-Way Switch Type", 1, 0, threeWaySwitchTypeOptions)
 }
@@ -1177,19 +1217,19 @@ Map getNightLightParam() {
 	return getParam(22, "Night Light Brightness", 1, 20, options)
 }
 
-// ZEN71/72/76/77
+// ZEN7x Models
 Map getLedColorParam() {
 	Integer num = 23
 	return getParam(num, "LED Indicator Color", 1, 1, ledColorOptions)
 }
 
-// ZEN71/72/76/77
+// ZEN7x Models
 Map getLedBrightnessParam() {
 	Integer num = 24
 	return getParam(num, "LED Indicator Brightness", 1, 1, ledBrightnessOptions)
 }
 
-// ZEN72/77 Only
+// ZEN7x Dimmers Only
 Map getSingleTapParam() {
 	return getParam(25, "Single Tap Up Brightness", 1, 0, singleTapOptions)
 }